translator_project/requirements.txt
streamlit
openai-whisper
transformers
sentencepiece
torch
gTTS
soundfile
